+++ src/fclaw_patch.h
@@ -1,0 +1,61 @@
+/*
+Copyright (c) 2012 Carsten Burstedde, Donna Calhoun
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+ * Redistributions of source code must retain the above copyright notice, this
+list of conditions and the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright notice,
+this list of conditions and the following disclaimer in the documentation
+and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+/** 
+ * @file
+ * Dimension-independent wrapper of patch
+ */
+
+#ifndef FCLAW_PATCH_H
+#define FCLAW_PATCH_H
+
+#include <fclaw_base.h>
+
+typedef struct fclaw_patch_user
+{
+    /** User defined patch structure */
+    void *user_patch;
+
+    /** Additional user data */
+    void *user_data;
+}
+fclaw_patch_user_t;
+
+typedef struct fclaw_patch
+{
+    union
+    {
+        /* avoid including dimension-specific files */
+        struct fclaw2d_patch_data *pd2;
+        struct fclaw3d_patch_data *pd3;
+    }
+    pd;
+
+    /* Those below are currently also in pd2, pd3.
+       Shall be removed from pd2, pd3. */
+    fclaw_patch_user_t pu;
+}
+fclaw_patch_t;
+
+#endif /* !FCLAW_PATCH_H */
